+++ CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to Merlin will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.7.5]
 
 ### Changed
-- Further improvements to the `null_spec` example.
 - Now requiring Celery version 5.x.
+- Further improvements to the `null_spec` example.
 
 ## [1.7.4]